requirements.txt
requests
python-dotenv
pandas
langchain
langchain-openai
langchain-experimental
tiktoken
numpy
chromadb
openai
cohere